base-files: Change the way base-files is modified to pass YP compat check
The base-files was being modified to always include the meta-measured fstab.

This would cause a project, where meta-measured was included, but IMA was
not enabled to always pick up the change.  Instead we want to check for the
ima distribution flag and only make the change if that is enabled.

An additional change to applying the securityfs to the already in place
fstab was made as it allows other recipes or user configurations to override
the fstab, but still ensure that we get the necessary securityfs in place
if needed.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>
(cherry picked from commit fbc288bce5576c63936f960822ffe366391b8c1d)
Signed-off-by: Ioan-Adrian Ratiu <adrian.ratiu@ni.com>